Refactor `type_util`
