+++ lib/streamlit/elements/lib/column_config_utils.py
@@ -21,10 +21,10 @@ from typing import TYPE_CHECKING, Dict, Final, Literal, Mapping, Union
 
 from typing_extensions import TypeAlias
 
+from streamlit.dataframe_util import DataFormat, is_colum_type_arrow_incompatible
 from streamlit.elements.lib.column_types import ColumnConfig, ColumnType
 from streamlit.elements.lib.dicttools import remove_none_values
 from streamlit.errors import StreamlitAPIException
-from streamlit.type_util import DataFormat, is_colum_type_arrow_incompatible
 
 if TYPE_CHECKING:
     import pyarrow as pa
