+++ lib/streamlit/runtime/state/session_state.py
@@ -40,13 +40,14 @@ from streamlit.proto.WidgetStates_pb2 import WidgetStates as WidgetStatesProto
 from streamlit.runtime.state.common import (
     RegisterWidgetResult,
     T,
+    ValueFieldName,
     WidgetMetadata,
+    is_array_value_field_name,
     is_keyed_widget_id,
     is_widget_id,
 )
 from streamlit.runtime.state.query_params import QueryParams
 from streamlit.runtime.stats import CacheStat, CacheStatsProvider, group_stats
-from streamlit.type_util import ValueFieldName, is_array_value_field_name
 
 if TYPE_CHECKING:
     from streamlit.runtime.session_manager import SessionManager
