+++ lib/streamlit/runtime/state/widgets.py
@@ -25,6 +25,7 @@ from streamlit.proto.WidgetStates_pb2 import WidgetState, WidgetStates
 from streamlit.runtime.state.common import (
     RegisterWidgetResult,
     T,
+    ValueFieldName,
     WidgetArgs,
     WidgetCallback,
     WidgetDeserializer,
@@ -37,7 +38,6 @@ from streamlit.runtime.state.common import (
 
 if TYPE_CHECKING:
     from streamlit.runtime.scriptrunner import ScriptRunContext
-    from streamlit.type_util import ValueFieldName
 
 
 ElementType: TypeAlias = str
